% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedpeToRearrCatalogue.R
\name{bedpeToRearrCatalogue}
\alias{bedpeToRearrCatalogue}
\title{bedpeToRearrCatalogue}
\usage{
bedpeToRearrCatalogue(sv_bedpe)
}
\arguments{
\item{sv_bedpe}{data frame BEDPE as described above}
}
\value{
returns the rearrangement catalogue for the given sample
}
\description{
This function converts a data frame BEDPE into a rearrangement catalogue, 
you should pass rearrangements of only one sample.
}
\details{
Columns present in the BEDPE should be:

"chrom1"

"start1"

"end1"

"chrom2"

"start2"

"end2"

"sample" sample name

"name" name of mate1

"partner" name of mate2

"score"

In addition, either the strands of the mates:

"strand1": + or -

"strand2": + or -

or the structural variant class

"svclass": translocation, inversion, deletion, tandem-duplication

According to Sanger notation you should have sv class equal to: (strand1/strand2)

inversion (+/-), if mates on the same chromosome

inversion (-/+), if mates on the same chromosome

deletion (+/+), if mates on the same chromosome

tandem-duplication (-/-), if mates on the same chromosome

translocation, if mates are on different chromosomes
}
\examples{
vcf_sv_file.bedpe <- "sample.bedpe"
sv_bedpe <- read.table(vcf_sv_file.bedpe,sep = "\\t",header = TRUE,
                     stringsAsFactors = FALSE,check.names = FALSE)
#build a catalogue from the bedpe file
res.cat <- bedpeToRearrCatalogue(sv_bedpe)
}
\keyword{bedpe}
\keyword{rearrangement}
