% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.GenomicScars.R
\name{calc.ai}
\alias{calc.ai}
\title{Number of Telomeric Allelic Imbalances}
\usage{
calc.ai(
  seg,
  chrominfo = chrominfo,
  min.size = 0,
  min.probes = 500,
  cont = 0,
  check.names = FALSE,
  ploidyByChromosome = TRUE,
  return.loc = FALSE,
  shrink = TRUE
)
}
\arguments{
\item{seg}{input data frame. Segmented output in the form of an ASCAT out matrix, which also includes ploidy and contamination. Total CN in column 6, nA in column 7, nB in column 8, ploidy in column 9, and contamination in column 10. Requires the following columns in the given order: "SampleID", "Chromosome", "Start", "End", "nProbes", "totalCN", "nA", "nB", "Ploidy" and    "AberrantCellFraction".}

\item{chrominfo}{data frame that contains necessary chromosome information. No need to specify this for human chromosomes, it will be loaded internally. A 3 column matrix with information about the chromosomes: chromosome name, chromosome length, centromere location.}

\item{min.size}{minimum size of segments}

\item{min.probes}{minimum number of probes in segments (I use 500 for the 900,000 probe SNP6, then scale down)}

\item{cont}{contamination threshold. By default set at 0 to ignore contamination}

\item{check.names}{check and potentially fix if there are duplicated samples. Any duplicates shall be re-named. Works by combining sample names with ploidy and SNP file name (which for this to work must be in column 11). Then looks for individual sample names which matches multiple original SNP files.}

\item{shrink}{joins segments of identical allelic copy number}
}
\description{
This function computes the number of Telomeric AI (NtAI). Implementation of Nicolai Juul Birkbak, njuul@cbs.dtu.dk (2014-07-02).
This functions generally rely on ASCAT or similarly processed copy number data, in a matrix format with at least the follwing columns in this exact order:
"SampleID", "Chromosome", "Start", "End", "nProbes", "totalCN", "nA", "nB", "Ploidy" and	"AberrantCellFraction".
NOTE: currently the chrominfo data refers to hg19, while hg38 is not yet supported.
}
\examples{
res <- calc.ai(seg)
}
\references{
Birkbak, N. J., Wang, Z. C., Kim, J.-Y., Eklund, A. C., Li, Q., Tian, R., ... Richardson, A. L. (2012). Telomeric Allelic Imbalance Indicates Defective DNA Repair and Sensitivity to DNA-Damaging Agents. Cancer Discovery, 2(4), 366–375. https://doi.org/10.1158/2159-8290.CD-11-0206
}
\author{
Nicolai Juul Birkbak, njuul@cbs.dtu.dk
}
