% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.GenomicScars.R
\name{calc.hrd}
\alias{calc.hrd}
\title{HRD-LOH index}
\usage{
calc.hrd(seg, nA = 7, check.names = FALSE, return.loc = FALSE)
}
\arguments{
\item{seg}{seg must be an ASCAT output object, in DNAcopy format. Requires the following columns in the given order: "SampleID", "Chromosome", "Start", "End", "nProbes", "totalCN", "nA", "nB", "Ploidy" and    "AberrantCellFraction". Column "nProbes", "Ploidy" and    "AberrantCellFraction" are not used here and can be set to NA.}

\item{nA}{is the column where copy number of A allele is found. This needs to be set to 7, because of the columns are referenced by position according to the required column order of seg.}
}
\description{
This function is an implementation of the cisplatin predictor developed by Myriad with Gordon Mills PMID: 23047548.
Abkevichs cutoffs: > 10, found in the supplementary info.
}
\examples{
HRD_LOH <- calc.hrd(ascat.data2, nA=7,check.names=FALSE, return.loc=FALSE)
}
\references{
Abkevich, V., Timms, K. M., Hennessy, B. T., Potter, J., Carey, M. S., Meyer, L. a., ... Lanchbury, J. S. (2012). Patterns of genomic loss of heterozygosity predict homologous recombination repair defects in epithelial ovarian cancer. British Journal of Cancer, 107(10), 1776–82. https://doi.org/10.1038/bjc.2012.451
}
\author{
Nicolai Juul Birkbak, njuul@cbs.dtu.dk
}
