% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.GenomicScars.R
\name{calc.lst}
\alias{calc.lst}
\title{Large-scale state transitions}
\usage{
calc.lst(
  seg,
  chrominfo = chrominfo.snp6,
  nA = 7,
  check.names = FALSE,
  return.loc = FALSE,
  chr.arm = "no"
)
}
\arguments{
\item{seg}{seg must be an ASCAT output object, in DNAcopy format. Requires the following columns in the given order: "SampleID", "Chromosome", "Start", "End", "nProbes", "totalCN", "nA", "nB", "Ploidy" and    "AberrantCellFraction". Column "nProbes" is not used here and can be set to NA.}

\item{nA}{is the column where copy number of A allele is found. This needs to be set to 7, because of the columns are referenced by position according to the required column order of seg.}
}
\description{
This function implements Popova's LST measure as described (Popova 2012, Cancer research)
Popova's cutoffs: 15 LSTs in near-diploid, 20 in near-tetraploid.
This functions generally rely on ASCAT or similarly processed copy number data, in a matrix format with at least the follwing columns in this exact order:
"SampleID", "Chromosome", "Start", "End", "nProbes", "totalCN", "nA", "nB", "Ploidy" and	"AberrantCellFraction".
NOTE: currently the chrominfo data refers to hg19, while hg38 is not yet supported.
}
\examples{
res <- calc.lst(seg)
}
\references{
Popova, T., Manié, E., Rieunier, G., Caux-Moncoutier, V., Tirapo, C., Dubois, T., ... Stern, M. H. (2012). Ploidy and large-scale genomic instability consistently identify basal-like breast carcinomas with BRCA1/2 inactivation. Cancer Research, 72(21), 5454–5462. https://doi.org/10.1158/0008-5472.CAN-12-1470
}
\author{
Nicolai Juul Birkbak, njuul@cbs.dtu.dk
}
