% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cataloguesClustering.R
\name{cataloguesClustering}
\alias{cataloguesClustering}
\title{Clustering of Catalogues or Signatures}
\usage{
cataloguesClustering(samplescatalogues, nclusters, outdir = NULL)
}
\arguments{
\item{samplescatalogues}{catalogues or signatures, with colnames as sample or signature names and rownames as the channel names}

\item{nclusters}{number of clusters, can be a single positive integer or a list of integers}

\item{outdir}{if specified, plot results into the outdir folder}
}
\value{
returns a list containing clustering results for each number of clusters requested
}
\description{
Cluster sample catalogues or signatures using hierarchical clustering with average linkage. The distance metric used is 1 - cosine similarity, so the clustering is based on the shape/direction of the catalogue/signature vectors.
}
\examples{
resClust <- cataloguesClustering(catalogues,2:10,"results/")
}
\keyword{clustering}
