% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandBiasLib.R
\name{combineStrandBiasResults}
\alias{combineStrandBiasResults}
\title{combine Transcription and Replication strand bias results}
\usage{
combineStrandBiasResults(biasResObjList)
}
\arguments{
\item{biasResObjList}{R list object containing a list of strand bias result objects, each obtained using the using the sampleStrandBias function. This should be a named list, for example each element name could be a sample name.}
}
\description{
This function combines the transcription/replication strand bias results obtained
using the sampleStrandBias function. For example, sampleStrandBias can be run multiple times
and each result object added to a named list, and then the list used as input for
the combineStrandBiasResults function. Results from multiple samples will then be
combined to obtain statistics such as average, standard deviation and p-value to
determine whether the bias is significant.
}
\examples{
biasResObjList <- list()
biasResObjList[["sample1"]] <- sampleStrandBias(sample1_snvs)
biasResObjList[["sample2"]] <- sampleStrandBias(sample2_snvs)
biasResObjList[["sample3"]] <- sampleStrandBias(sample3_snvs)
res <- combineStrandBiasResults(biasResObjList)
}
