% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{computeCorrelation_parallel}
\alias{computeCorrelation_parallel}
\title{Compute Correlation (parallel)}
\usage{
computeCorrelation_parallel(x, nparallel = 1, parallel = FALSE)
}
\arguments{
\item{x}{catalogues/signature matrix where each column is a catalogue/signature and each row is a channel.}

\item{nparallel}{how many parallel processes to use}

\item{parallel}{set to TRUE in order to use parallel}
}
\value{
correlation matrix
}
\description{
Compute the correlation of a set of signatures/catalogues according to cosine similarity
}
