% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{computeCorrelationOfTwoSetsOfSigs}
\alias{computeCorrelationOfTwoSetsOfSigs}
\title{computeCorrelationOfTwoSetsOfSigs}
\usage{
computeCorrelationOfTwoSetsOfSigs(sigs1, sigs2)
}
\arguments{
\item{sig1}{matrix of signatures, with signatures as columns}

\item{sig2}{matrix of signatures, with signatures as columns}
}
\value{
cosine similarity matrix
}
\description{
Compute the cosine similarity between two sets of signatures, which results in a cosine similarity matrix.
}
