% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DinucleotideVariantsLib.R
\name{convertToAlexandrovChannels}
\alias{convertToAlexandrovChannels}
\title{Convert DNV catalogues from Zou's to Alexandrov's style}
\usage{
convertToAlexandrovChannels(dnvCatalogues)
}
\arguments{
\item{dnvCatalogues}{DNV catalogues in Zou's style}
}
\value{
DNV Catalogues in Alexandrov style
}
\description{
Function to convert DNV signatures or catalogues from Zou's to Alexandrov's style. This will simply move and rename the channels.
}
