% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{cos_sim}
\alias{cos_sim}
\title{Cosine Similarity}
\usage{
cos_sim(a, b)
}
\arguments{
\item{a}{first vector to compare}

\item{b}{second vector to compare}
}
\value{
cosine similarity
}
\description{
Compute the cosine similarity between two vectors, using the formula sum(a*b)/sqrt(sum(a^2)*sum(b^2)).
}
