% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceEvaluationLib.R
\name{evaluatePerformanceExposures}
\alias{evaluatePerformanceExposures}
\title{evaluate performance exposures}
\usage{
evaluatePerformanceExposures(
  true_exposures,
  estimated_exposures,
  commonNames = NULL,
  rareNames = NULL,
  outfile = NULL
)
}
\arguments{
\item{true_exposures}{matrix of true exposures, for example obtained using simulations, with mutational signatures names as colnames and sample names as rownames}

\item{estimated_exposures}{matrix of estimated exposures, for example obtained using a signature extraction or fit algorithm on a simulated dataset, with mutational signatures names as colnames and sample names as rownames}

\item{commonNames}{names of common mutational signatures. If NULL, all signatures in the colnames of true_exposures and estimated_exposures will be used}

\item{rareNames}{names of rare mutational signatures. If NULL, rare signatures metrics will not be computed. This is ok if the simulated data didn't have any rare signatures.}

\item{outfile}{file name for plotting, please use .pdf file name extension. Can be omitted.}
}
\value{
performance metrics
}
\description{
Compare two exposure matrices, one with the true exposures and one with estimated exposures, and
compute various metrics to determine how well the estimated exposures match the true exposures.
If the names of common and/or rare mutational signatures are provided, then the
metrics are calculated separately for common and rare signatures.
Results are plotted to a file if outfile is specified.
}
