% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceEvaluationLib.R
\name{evaluatePerformanceExposuresList}
\alias{evaluatePerformanceExposuresList}
\title{evaluate performance exposures list}
\usage{
evaluatePerformanceExposuresList(
  true_exposures,
  estimated_exposures_list,
  commonNames = NULL,
  rareNames = NULL,
  outfile = NULL
)
}
\arguments{
\item{true_exposures}{matrix of true exposures, for example obtained using simulations, with mutational signatures names as colnames and sample names as rownames}

\item{estimated_exposures_list}{list of matrices of estimated exposures, for example obtained using a signature extraction or fit algorithm on a simulated dataset, with mutational signatures names as colnames and sample names as rownames. Use a named list, so that the names can be used in he data structure returned and plots.}

\item{commonNames}{names of common mutational signatures. If NULL, all signatures in the colnames of true_exposures and estimated_exposures will be used}

\item{rareNames}{names of rare mutational signatures. If NULL, rare signatures metrics will not be computed. This is ok if the simulated data didn't have any rare signatures.}

\item{outfile}{file name for plotting, please use .pdf file name extension. Can be omitted.}
}
\value{
performance metrics
}
\description{
Compare an exposure matrix with the true exposures with a list of estimated exposure matrices, and
compute various metrics to determine how well the estimated exposures match the true exposures.
If the names of common and/or rare mutational signatures are provided, then the
metrics are calculated separately for common and rare signatures.
Results are plotted to a file if outfile is specified.
}
