% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignMutationsSignatureProbability.R
\name{expandColumnToMatrix}
\alias{expandColumnToMatrix}
\title{Expand columns to matrix}
\usage{
expandColumnToMatrix(dataTable, targetColname)
}
\arguments{
\item{dataTable}{table containing a column to expand. Each cell to expand will be a list of pairs, where each element of the list is separated by ";" and each pair is separated by ":". Pairs indicate colname and value. For example, if the value in a row is C1:V1;C2:V2;C4:V4, this will be expanded into columns C1, C2, C4 and values for that row V1, V2, V4.}

\item{targetColname}{colname of the column to expand}
}
\value{
dataTable with expanded column
}
\description{
Expand columns to matrix
}
