% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFitLib.R
\name{export_SignatureFit_withBootstrap_to_JSON}
\alias{export_SignatureFit_withBootstrap_to_JSON}
\title{Export Signature Fit with bootstrap to JSON}
\usage{
export_SignatureFit_withBootstrap_to_JSON(outdir, res)
}
\arguments{
\item{outdir}{output directory where the output JSON files should be saved, remember to put "/" at the end of the folder name.}

\item{res}{R object obtained from the SignatureFit_withBootstrap or SignatureFit_withBootstrap_Analysis function}
}
\description{
Given a res file obtained from the SignatureFit_withBootstrap or 
SignatureFit_withBootstrap_Analysis function, export it to a set of 
JSON files that can be used for web visualisation
}
\examples{
res <- SignatureFit_withBootstrap(cat = rnd_matrix,
                  signature_data_matrix = cosmic30,
                  nboot = 5,
                  threshold_percent = 0.1,
                  threshold_p.value = 0.1)
export_SignatureFit_withBootstrap_to_JSON("JSON_out/",res)
}
\keyword{JSON}
\keyword{fit}
\keyword{mutational}
\keyword{signatures}
