% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{findClosestCOSMIC30}
\alias{findClosestCOSMIC30}
\title{Find closest COSMIC30 signatures}
\usage{
findClosestCOSMIC30(sigs, threshold)
}
\arguments{
\item{sigs}{matrix with 96-channel substitution signatures as columns}

\item{threshold}{cosine similarity threshold}
}
\value{
the list of signatures identified
}
\description{
Compares a set of signatures to the COSMIC30 signatures and
returns the list of signatures identified. For example,
c("C1","C3","N1","C13","N2") means that Cosmic (C) signatures 1, 3 and 13 were found, while
signatures N1 and N2 are unknown signatures (N for not found), based on a similarity threshold (similarity >=threshold)
comparing to COSMIC30
}
