% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{findClosestCOSMIC30_withSimilarity}
\alias{findClosestCOSMIC30_withSimilarity}
\title{Find closest COSMIC30 signatures}
\usage{
findClosestCOSMIC30_withSimilarity(sigs)
}
\arguments{
\item{sigs}{matrix with 96-channel substitution signatures as columns}
}
\value{
the list of signatures identified and corresponding similarities
}
\description{
Compares a set of signatures to the COSMIC30 signatures and
returns the list of signatures identified and the corresponding similarity. For example,
list(cosmic = c("C1","C3","C13"),cos_sim = c(0.94,0.85,0.7))
means that Cosmic (C) signatures 1, 3 and 13 were found, while
the corrsponding similarities to those signatures are 0.94, 0.85 and 0.7
}
