% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationContext.R
\name{findContextSNV}
\alias{findContextSNV}
\title{Find SNV context}
\usage{
findContextSNV(
  snv_table,
  mtype = NULL,
  fiveprime = NULL,
  threeprime = NULL,
  genomev = "hg19",
  context_length = 5
)
}
\arguments{
\item{snv_table}{data frame containing SNVs, with required columns chr, position, REF, ALT}

\item{mtype}{if specified, filter mutations according to specific base changes: C>A, C>T, C>G, T>A, T>C, T>G}

\item{fiveprime}{if specified, filter mutations according to specific 3': C, A, T, G. Used only if mtype is specified}

\item{genomev}{genome version, hg19 or hg38}

\item{context_length}{number of nucleotides on the 5' and 3' to consider}
}
\value{
table of counts of base contexts
}
\description{
Given a table of SNVs, find the reference genome nucleotide bases that are at
the 5' and 3' of each SNV location, according to the pyrimidine reference. For
example, if the mutation is a TT[C>T]AG, then the 5' will correspond to lower genomic
coordinates, with 5' context TT and the 3' will correspond to the higher genomic coordinates,
with 3' context AG. Alternatively, if the mutation is a TT[A>T]AG, then the reference
considered will be on the opposite strand, and the mutation and contexts will be
reversed as CT[T>A]AA.
It is also possible to filter the mutations in the given SNV table according to
the specific pyrimidine mutation and, optionally, the first base on the 5' or 3'.
}
