% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findKataegis.R
\name{findKataegis}
\alias{findKataegis}
\title{find kataegis}
\usage{
findKataegis(snvs_table, sample_name)
}
\arguments{
\item{snvs_table}{table listing the single nucleotide variants to use. Necessary column names are: chr, position, REF, ALT.}

\item{sample_name}{name of sample}
}
\value{
kataegis regions and annotated SNVs
}
\description{
Given a table with a list of single nucleotide variants (SNVs) this function
identifies groups of SNVs that are narrowly clustered together, and that are
likely part of a kataegis. An annotated table of SNVs and a list of kataegis
regions are returned.
}
