% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{Fit}
\alias{Fit}
\title{Signature Fit}
\usage{
Fit(
  catalogues,
  signatures,
  exposureFilterType = "fixedThreshold",
  giniThresholdScaling = 10,
  threshold_percent = 5,
  method = "KLD",
  useBootstrap = FALSE,
  nboot = 200,
  threshold_p.value = 0.05,
  nparallel = 1,
  randomSeed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{catalogues}{catalogues matrix, samples as columns, channels as rows}

\item{signatures}{mutational signatures to bw fitted into the sample catalgues, signatures as columns and channels as rows}

\item{exposureFilterType}{use either fixedThreshold or giniScaledThreshold. When using fixedThreshold, exposures will be removed based on a fixed percentage with respect to the total number of mutations (threshold_percent will be used). When using giniScaledThreshold each signature will used a different threshold calculated as (1-Gini(signature))*giniThresholdScaling}

\item{giniThresholdScaling}{scaling factor for the threshold type giniScaledThreshold, which is based on the Gini score of a signature}

\item{threshold_percent}{threshold in percentage of total mutations in a sample, only exposures larger than threshold are considered}

\item{method}{KLD or NNLS}

\item{useBootstrap}{set to TRUE to use the signature fit with bootstrap method}

\item{nboot}{number of bootstraps to use, more bootstraps more accurate results (use only when useBootstrap=TRUE)}

\item{threshold_p.value}{p-value to determine whether an exposure is above the threshold_percent.
In other words, this is the empirical probability that the exposure is lower than the threshold (use only when useBootstrap=TRUE)}

\item{nparallel}{to use parallel specify >1}

\item{randomSeed}{set an integer random seed (use only when useBootstrap=TRUE)}

\item{verbose}{use FALSE to suppress messages}
}
\value{
returns the activities/exposures of the signatures in the given sample and other information
}
\description{
This function provides basic signature fit functionalities.
Fit a given set of mutational signatures into mutational catalogues to estimate
the activty/exposure of each of the given signatures in the catalogues.
}
\details{
This is a standard interface to signature fit functions with/without bootstrap. The object returned by this
function can be passed to the plotFit() function for automated plotting of the results.

A post fit exposure filter will reduce the false positive singature assignments by setting to zero exposure values that
are below a certain threshold. We provide two exposureFilterType methods: fixedThreshold and giniScaledThreshold. The
fixedThreshold method will set to zero exposures that are below a fixed threshold given as a percentage of the mutations
in a sample (parameter threshold_percent), while the method giniScaledThreshold will use a different threshold for each
signature, computed as (1-Gini(signature))*giniThresholdScaling, which will also be a percentage of the mutations in a sample.
}
\examples{
res <- Fit(catalogues,getOrganSignatures("Breast"))
plotFit(res,"results/")
}
\keyword{fit}
\keyword{mutational}
\keyword{signatures}
