% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{FitMS}
\alias{FitMS}
\title{Multi-Step Mutational Signatures Fit}
\usage{
FitMS(
  catalogues,
  organ = NULL,
  rareSignatureTier = "T2",
  commonSignatures = NULL,
  rareSignatures = NULL,
  method = "KLD",
  exposureFilterType = "fixedThreshold",
  threshold_percent = 5,
  giniThresholdScaling = 10,
  multiStepMode = "errorReduction",
  residualNegativeProp = 0.003,
  minResidualMutations = NULL,
  minCosSimRareSig = 0.8,
  minErrorReductionPerc = 15,
  minCosSimIncrease = 0.02,
  useBootstrap = FALSE,
  nboot = 200,
  threshold_p.value = 0.05,
  maxRareSigsPerSample = 1,
  nparallel = 1,
  randomSeed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{catalogues}{catalogues matrix, samples as columns, channels as rows}

\item{organ}{#automatically sets the commonSignatures and rareSignatures parameters, which can be left as NULL. The following organs are available:
"Biliary", "Bladder", "Bone_SoftTissue", "Breast", "CNS", "Colorectal", "Esophagus", "Head_neck", "Kidney", "Liver", "Lung", "Lymphoid", "Myeloid",
"NET", "Oral_Oropharyngeal", "Ovary", "Pancreas", "Prostate", "Skin", "Stomach", "Uterus"}

\item{rareSignatureTier}{either T1 or T2. For each organ we provide two lists of rare signatures that can be used. Tier 1 (T1) are rare signatures
that were observed in the requested organ. The problem with T1 is that it may be that a signature is not observed simply because there were not enough samples for a certain organ in the particular
dataset that was used to extract the signatures. So in general we advise to use Tier 2 (T2) signatures, which extend the rare signature to a wider number of rare signatures.}

\item{commonSignatures}{signatures, signatures as columns, channels as rows. These are the signatures that are assumed to be present in most samples and will be used in the first step.
Can be set automatically by specifying the organ parameter}

\item{rareSignatures}{signatures, signatures as columns, channels as rows. These are the signatures that are assumed to be rarely present in a sample, at most maxRareSigsPerSample rare signatures in each sample.
Can be set automatically by specifying the organ parameter and the rareSignatureTier parameter}

\item{method}{KLD or NNLS}

\item{exposureFilterType}{use either fixedThreshold or giniScaledThreshold. When using fixedThreshold, exposures will be removed based on a fixed percentage with respect to the total number of mutations (threshold_percent will be used). When using giniScaledThreshold each signature will used a different threshold calculated as (1-Gini(signature))*giniThresholdScaling}

\item{threshold_percent}{threshold in percentage of total mutations in a sample, only exposures larger than threshold are considered}

\item{giniThresholdScaling}{scaling factor for the threshold type giniScaledThreshold, which is based on the Gini score of a signature}

\item{multiStepMode}{use one of the following: "constrainedFit", "partialNMF", "errorReduction", or "cossimIncrease".}

\item{residualNegativeProp}{maximum proportion of mutations (w.r.t. total mutations in a sample) that can be in the negative part of a residual when using the constrained least squares fit
when using multiStepMode=constrainedFit}

\item{minResidualMutations}{minimum number of mutations in a residual when using constrainedFit or partialNMF. Deactivated by default.}

\item{minCosSimRareSig}{minimum cosine similarity between a residual and a rare signature for considering the rare signature as a candidate for a sample when using constrainedFit or partialNMF}

\item{minErrorReductionPerc}{minimum percentage of error reduction for a signature to be considered as candidate when using the errorReduction method. The error is computed as mean absolute deviation}

\item{minCosSimIncrease}{minimum cosine similarity increase for a signature to be considered as candidate when using the cossimIncrease method}

\item{useBootstrap}{set to TRUE to use bootstrap}

\item{nboot}{number of bootstraps to use, more bootstraps more accurate results}

\item{threshold_p.value}{p-value to determine whether an exposure is above the threshold_percent. In other words, this is the empirical probability that the exposure is lower than the threshold}

\item{maxRareSigsPerSample}{masimum number of rare signatures that should be serched in each sample. In most situations, leaving this at 1 should be enough.}

\item{nparallel}{to use parallel specify >1}

\item{randomSeed}{set an integer random seed}

\item{verbose}{use FALSE to suppress messages}
}
\value{
returns the activities/exposures of the signatures in the given sample and other information
}
\description{
Given a set of mutational catalogues, this function will attempt fit mutational signature in a multi-step manner. In the first step, only the common signatures are fitted into the samples.
In the following steps, one or more rare signatures are fitted into the samples in addition to the common signatures. Common and rare signatures can be determined automatically by providing
the name of an organ, or can be supplied by the user.
}
\details{
We provide four methods to identify the rare signatures in the samples:
"constrainedFit", "partialNMF", "errorReduction", or "cossimIncrease". The methods constrainedFit and partialNMF work in a similar way:
they identify a residual in each given sample, as the leftover mutations after fitting the common signatures. They will attempt to produce a mostly positive residual.
Each residual is then compared to each rare signature, and a rare signature is considered as a candidate rare signature for a sample if the cosine similarity between the residual
and the signature is at least minCosSimRareSig. One can also request that the residual is at least minResidualMutations. While constrainedFit will use a constrained least square fit where
the negative part of the residual is at most residualNegativeProp (a proportion of the number of mutations in the sample), partialNMF will instead use a few iterations of a KLD based NMF algorithm
where the matrix of the signatures contains the common signature and an additional signatures that needs to be estimated (NNLM package). The methods errorReduction and cossimIncrease work
in a similar way: they will fit the common signatures along with one additional rare signature, testing all rare signatures one at a time, and then determine difference in error
(or cosine similarity) between fitting the common signatures only and with the additional rare signatures. If the error reduction is at least minErrorReductionPerc (or the cosine similarity
increase is at least minCosSimIncrease then the rare signature will be considered as a candidate.

After any of ghe procedures above, each sample may have multiple candidate rare signatures, so one is chosen according to the highest associated cosine similarity either of
the residual to the candidate rare signature (constrainedFit and partialNMF methods), or of the catalogue and the reconstructed sample (errorReduction and cossimIncrease methods).
It is then possible to plot all the fits with plotFitMS and even change the choise of the candidate rare signature using the function fitMerge.

A post fit exposure filter will reduce the false positive singature assignments by setting to zero exposure values that
are below a certain threshold. We provide two exposureFilterType methods: fixedThreshold and giniScaledThreshold. The
fixedThreshold method will set to zero exposures that are below a fixed threshold given as a percentage of the mutations
in a sample (parameter threshold_percent), while the method giniScaledThreshold will use a different threshold for each
signature, computed as (1-Gini(signature))*giniThresholdScaling, which will also be a percentage of the mutations in a sample.
}
\examples{
res <- FitMS(catalogues,"Breast")
plotFitMS(res,"results/")
}
\references{
A. Degasperi, X. Zou, T. D. Amarante, ..., H. Davies, Genomics England Research Consortium, S. Nik-Zainal. Substitution mutational signatures in whole-genome-sequenced cancers in the UK population. Science, 2022.
}
\keyword{fit}
\keyword{mutational}
\keyword{signatures}
