% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{fitMStoJSON}
\alias{fitMStoJSON}
\title{Export the results from the FitMS function to JSON}
\usage{
fitMStoJSON(
  fitObj,
  filename = "export.json",
  nindent = 0,
  disableFileWrite = FALSE,
  blockname = NULL
)
}
\arguments{
\item{fitObj}{object obtained from the FitMS function}

\item{filename}{file name where to save the JSON string}

\item{nindent}{number of tabs of indentation to be used, default=0. This is useful in case one wants to embed the output in a larger JSON document}

\item{disableFileWrite}{if TRUE the JSON lines are simply printed to screen. This option is used when fitToJSON is called from another function}

\item{blockname}{if specified, the first line will show the given name as the block name, which is useful if the output is embedded in a larger JSON file}
}
\description{
Exports of the results obtained with the FitMS function to JSON.
}
\examples{
res <- FitMS(catalogues,"Breast")
fitMStoJSON(res,"export.json")
}
