% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureIdentificationLib.R
\name{fitSignaturesWithSignatures}
\alias{fitSignaturesWithSignatures}
\title{fitSignaturesWithSignatures}
\usage{
fitSignaturesWithSignatures(
  signatures_to_fit,
  signatures_to_use,
  max_combinations = 2,
  sig_min_contrib = 0.15,
  minsim = 0.75,
  max_from_each_class = 10,
  nparallel = 1
)
}
\arguments{
\item{signatures_to_fit}{matrix with the mutational signatures that need to be identified/fitted. Each column has a signature and each row corresponds to a mutational channel. Columns need to sum to 1.}

\item{signatures_to_use}{matrix of mutational signatures to be used as a reference and that will be fitted/combined to match the signatures_to_fit. Each column has a signature and each row corresponds to a mutational channel. Columns need to sum to 1.}

\item{max_combinations}{maximum number of signatures_to_use to be used as a linear combination to fit each signatures_to_fit}

\item{sig_min_contrib}{minimum % of contribution that a signatures_to_use should have to a linear combination when matching a signature in signatures_to_fit}

\item{minsim}{minimum cosine similarity between a signatures_to_fit and a match, where a match is either a signature in signatures_to_use or a linear combination of signatures in signatures_to_use}

\item{max_from_each_class}{maximum number of matches to consider in each class, where the class indicates the number of signatures in a linear combination. If there are more matches in the class than the maximum specified, then the matches with the highest cosine similarity will be reported}

\item{nparallel}{how many parallel cpus to use}
}
\value{
tables with the best linear combinations of signatures in signatures_to_use that match signatures in signatures_to_fit
}
\description{
This function attempts to reconstruct a set of signatures (signatures_to_fit) using
a second set of signatures (signatures_to use), either just checking the similarity
between the signatures of the first and the second set, or using a linear combination
of signatures in the second set to model the first set. The typical use of this
function would be to check whether a set of extracted signatures can be matched to
a set of known signatures, a procedure often called signature assignment or signature
identification.
}
