% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomePlot.R
\name{genomePlot}
\alias{genomePlot}
\title{Genome Plot}
\usage{
genomePlot(
  subsVcf.file,
  indelsVcf.file,
  cnvsTab.file,
  rearrBedpe.file,
  sampleID,
  genome.v = "hg19",
  ...,
  file.ideogram = NULL,
  plot_title = NULL,
  no_copynumber = FALSE,
  no_rearrangements = FALSE,
  no_indels = FALSE,
  no_subs_legend = FALSE,
  out_format = "png",
  out_path = ".",
  rearr_only_assembled = FALSE,
  base.per.unit = NULL
)
}
\arguments{
\item{subsVcf.file}{SNV VCF file. The file should only contain SNV and should already be filtered according to the user preference, as all SNV in the file will be used and no filter will be applied.}

\item{indelsVcf.file}{Indels VCF file to be used to classify Indels and compute the proportion of indels at micro-homology. The files should only contain indels (no SNV) and should already be filtered according to the user preference, as all indels in the file will be used and no filter will be applied.}

\item{cnvsTab.file}{CNV TAB file (similar to ASCAT format). The file should be tab separated and contain a header in the first line with the following columns: 'seg_no', 'Chromosome', 'chromStart', 'chromEnd', 'total.copy.number.inNormal', 'minor.copy.number.inNormal', 'total.copy.number.inTumour', 'minor.copy.number.inTumour'}

\item{rearrBedpe.file}{SV (Rearrangements) BEDPE file. The file should contain a rearrangement for each row (two breakpoint positions should be on one row as determined by a pair of mates of paired-end sequencing) and should already be filtered according to the user preference, as all rearrangements in the file will be used and no filter will be applied. The file should contain a header in the first line with the following columns: "chrom1", "start1", "end1", "chrom2", "start2", "end2" and "sample" (sample name). In addition, either two columns indicating the strands of the mates, "strand1" (+ or -) and "strand2" (+ or -), or one column indicating the structural variant class, "svclass": translocation, inversion, deletion, tandem-duplication. The column "svclass" should correspond to (Sanger BRASS convention): inversion (strands +/- or -/+ and mates on the same chromosome), deletion (strands +/+ and mates on the same chromosome), tandem-duplication (strands -/- and mates on the same chromosome), translocation (mates are on different chromosomes).}

\item{sampleID}{Name of the sample.}

\item{genome.v}{set genome version: hg19, hg38, mm10 or canFam3.}

\item{file.ideogram}{name of the file that contain a user defined genome ideogram. Leave to NULL to load appropriate ideogram according to genome version.}

\item{plot_title}{title of the plot.}

\item{no_copynumber}{set to TRUE to disable plotting of copy number data}

\item{no_rearrangements}{set to TRUE to disable plotting of rearrangement data}

\item{no_indels}{set to TRUE to disable plotting of indels data}

\item{no_subs_legend}{set to TRUE to disable plotting of substitutions legend}

\item{out_format}{set to either png or svg}

\item{out_path}{directory where to place the plot.}

\item{rearr_only_assembled}{only include the rearrangements that have an assembly score}

\item{base.per.unit}{set RCircos base.per.unit parameter. Useful for whole exome data.}
}
\value{
return the generated plot file name.
}
\description{
Generates a plot for the visualisation of somatic variants across the genome, organised in a circle.
Variants plotted are single nucleotide variations (SNV), small insertions and deletions (indels),
copy number variations (CNV) and rearrangements.
}
