% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{getCOSMICSignatures}
\alias{getCOSMICSignatures}
\title{getCOSMICSignatures}
\usage{
getCOSMICSignatures(version = "latest", typemut = "subs", verbose = TRUE)
}
\arguments{
\item{version}{this is either "latest", which is v3.2 for SBS and DBS, or it is possible to specify "2" in combination with typemut="subs" to obtain the old COSMIC 30 signatures. For rearrangements, this function only returns the 6 breast cancer rearr signatures from Nik-Zainal et al. 2016.}

\item{typemut}{either subs, DNV or rearr}
}
\value{
reference signatures matrix
}
\description{
This function returns the COSMIC signatures for a given mutation type.
}
