% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{getGiniThresholds}
\alias{getGiniThresholds}
\title{Get Gini Thresholds}
\usage{
getGiniThresholds(
  signatures,
  giniThresholdScaling,
  giniThresholdScaling_nmuts,
  verbose = TRUE
)
}
\arguments{
\item{signatures}{matrix of mutational signatures, with signatures as columns and channels as rows}

\item{giniThresholdScaling}{scaling factor for relative threshold (percentage of mutations)}

\item{giniThresholdScaling_nmuts}{scaling factor for absolute threshold (number of mutations)}
}
\description{
Given a matrix of mutational signatures and Gini threshold scaling factors,
return the value of the thresholds.
}
