% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{getOrganSignatures}
\alias{getOrganSignatures}
\title{getOrganSignatures}
\usage{
getOrganSignatures(organ, typemut = "subs")
}
\arguments{
\item{organ}{one of the following: "Biliary", "Bladder", "Bone_SoftTissue", "Breast", "Cervix", "CNS", "Colorectal", "Esophagus", "Head_neck", "Kidney", "Liver", "Lung", "Lymphoid", "Ovary", "Pancreas", "Prostate", "Skin", "Stomach", "Uterus"}

\item{typemut}{either subs or rearr}
}
\value{
organ-specific signatures matrix
}
\description{
This function returns the organ-specific signatures for a given organ and mutation type as defined in Degasperi et al. 2020 Nat Cancer paper.
}
\references{
A. Degasperi, T. D. Amarante, J. Czarnecki, S. Shooter, X. Zou, D. Glodzik, S. Morganella, A. S. Nanda, C. Badja, G. Koh, S. E. Momen, I. Georgakopoulos-Soares, J. M. L. Dias, J. Young, Y. Memari, H. Davies, S. Nik-Zainal. A practical framework and online tool for mutational signature analyses show intertissue variation and driver dependencies, Nature Cancer, https://doi.org/10.1038/s43018-020-0027-5, 2020.
}
