% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDetect.R
\name{HRDetect_pipeline}
\alias{HRDetect_pipeline}
\title{HRDetect Pipeline}
\usage{
HRDetect_pipeline(data_matrix, genome.v = "hg19", SNV_vcf_files = NULL,
  SNV_tab_files = NULL, SNV_catalogues = NULL, Indels_vcf_files = NULL,
  CNV_vcf_files = NULL, SV_bedpe_files = NULL, SV_catalogues = NULL,
  nparallel = 1)
}
\description{
Run the HRDetect pipeline
}
