% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{KLD}
\alias{KLD}
\title{KL-divergence}
\usage{
KLD(m1, m2)
}
\arguments{
\item{m1}{original matrix}

\item{m2}{matrix to be used to approximate m1}
}
\value{
KL-Divergence
}
\description{
Compute the Kullback-Leibler Divergence between two matrices. In order to compute the divergence, .Machine$double.eps is added to matrices zero entries.
}
