% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{loadFitFromFile}
\alias{loadFitFromFile}
\title{Load fit object from file}
\usage{
loadFitFromFile(filename, verbose = T)
}
\arguments{
\item{filename}{file name from which to load the fit object}
}
\description{
This function loads a Fit or FitMS object from an R data file.
In order to work, the file must have been generated using the saveFitToFile function,
which will save the fit object with a standard name.
}
\examples{
fitObj <- loadFitFromFile("fit.rData")
}
