% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCatalogueSimulationLib.R
\name{loadSimulatedData}
\alias{loadSimulatedData}
\title{Load simulated data}
\usage{
loadSimulatedData(simDir)
}
\arguments{
\item{simDir}{directory where the simulated data was saved}
}
\value{
simulated data object
}
\description{
Function that loads simulated data that was previously saved with the saveSimulatedData
function. The function can load  both data saved as R object and as plain text,
identifying the correct format automatically.
}
