% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchedClustering.R
\name{matchedClustering}
\alias{matchedClustering}
\title{Clustering with Matches}
\usage{
matchedClustering(
  distMatrix,
  ns,
  maxMatch = TRUE,
  parallel = FALSE,
  nparallel = 1
)
}
\arguments{
\item{distMatrix}{squared distance matrix}

\item{ns}{number of clusters}

\item{maxMatch}{if TRUE, use the maximum match (assignament problem), otherwise use a stable matching (which may not be maximum). Maximum means that the match will have minimal sum of distances}

\item{parallel}{use multi-threading TRUE/FALSE, uses doMC and foreach packages}

\item{nparallel}{how many threads at a time}
}
\value{
returns the clusters assigned to each element (there is an element for each column of distMatrix)
}
\description{
Compute the clustering into ns clusters of elements for which a distance matrix (distMatrix) is given, subject to the constraint that groups of ns elements cannot be inthe same cluster. For example, elements from 1 to ns cannot be in the same cluster, and so it is for elements from ns+1 to 2*ns, 2*ns+1 to 3*ns, and so on.
}
\examples{
res.clust <- matchedClustering(distMatrix,10)
}
\keyword{assignment}
\keyword{clustering}
\keyword{constraints}
\keyword{matching}
\keyword{problem}
\keyword{stable}
