% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCatalogueSimulationLib.R
\name{mutationSampling}
\alias{mutationSampling}
\title{mutationSampling}
\usage{
mutationSampling(allSignatures, exposures)
}
\arguments{
\item{allSignatures}{mutational signatures matrix with the signatures to sample}

\item{exposures}{exposures with how many muts for each signatures in each sample}
}
\value{
sampled catalogues
}
\description{
mutationSampling
}
