% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{plot.CosSimSignatures}
\alias{plot.CosSimSignatures}
\title{plot.CosSimSignatures}
\usage{
\method{plot}{CosSimSignatures}(sig1, sig2, output_file, dpi = 300,
  xlabel = "", ylabel = "")
}
\arguments{
\item{sig1}{matrix with signatures as columns}

\item{sig2}{matrix with signatures as columns}

\item{output_file}{name of the output file (jpg)}

\item{dpi}{dots per inch}

\item{xlabel}{label for x axis}

\item{ylabel}{label for y axis}
}
\description{
Plot a matrix of cosine similarities between two sets of signatures.
}
