% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDetect.R
\name{plot_HRDetect_BootstrapScores}
\alias{plot_HRDetect_BootstrapScores}
\title{plot_HRDetect_BootstrapScores}
\usage{
plot_HRDetect_BootstrapScores(outdir, hrdetect_res,
  main = "Distribution of HRDetect scores", mar = c(9, 4, 3, 2),
  pwidth = NULL, pheight = 1000, pres = 200)
}
\arguments{
\item{outdir}{output directory for the plot}

\item{hrdetect_res}{output of HRDetect_pipeline, ran with bootstrap_scores=TRUE}

\item{main}{plot title. If not specified: "Distribution of HRDetect scores"}

\item{mar}{plot margin. If not specified: c(9,4,3,2)}

\item{pwidth}{plot width in pixels. If not specified: max(1000,400+120*number_of_samples)}

\item{pheight}{plot height in pixels. If not specified: 1000}

\item{pres}{plot resolution. If not specified: 200}
}
\description{
Overall plot of scores obtained from HRDetect with bootstrap.
}
