% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDetect.R
\name{plot_HRDetect_Contributions}
\alias{plot_HRDetect_Contributions}
\title{plot_HRDetect_Contributions}
\usage{
plot_HRDetect_Contributions(
  file_name,
  hrdetect_output,
  par_title = "HRDetect Score and Contributions"
)
}
\arguments{
\item{file_name}{name of the output file (pdf)}

\item{hrdetect_output}{output of HRDetect, containing contributions of each feature}

\item{par_title}{title of the plot (optional)}
}
\description{
Function for plotting HRDetect results and contributions to HRDetect score.
}
