% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDetect.R
\name{plot_HRDetect_overall}
\alias{plot_HRDetect_overall}
\title{plot_HRDetect_overall}
\usage{
plot_HRDetect_overall(file_name, hrdetect_output)
}
\arguments{
\item{file_name}{name of the output file (jpg)}

\item{hrdetect_output}{output of HRDetect, containing contributions of each feature}
}
\description{
Overall plot of scores obtained from HRDetect.
}
