% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDetect.R
\name{plot_HRDLOH_HRDetect_Contributions}
\alias{plot_HRDLOH_HRDetect_Contributions}
\title{plot_HRDLOH_HRDetect_Contributions}
\usage{
plot_HRDLOH_HRDetect_Contributions(file_name, HRDLOH_index, hrdetect_output)
}
\arguments{
\item{file_name}{name of the output file (jpg)}

\item{HRDLOH_index}{HRD index score}

\item{hrdetect_output}{output of HRDetect, containing contributions of each feature}
}
\description{
Function for plotting HRDetect results and contributions to HRDetect score.
}
