% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFitLib.R
\name{plot_SignatureFit_withBootstrap}
\alias{plot_SignatureFit_withBootstrap}
\title{plot_SignatureFit_withBootstrap}
\usage{
plot_SignatureFit_withBootstrap(outdir, boostrapFit_res, type_of_mutations)
}
\arguments{
\item{outdir}{output directory for the plot}

\item{boostrapFit_res}{output of SignatureFit_withBootstrap}

\item{type_of_mutations}{type of mutations: subs, rearr, or generic}
}
\description{
Generate plots from the output of SignatureFit_withBootstrap.
}
