% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandBiasLib.R
\name{plotBiasResults}
\alias{plotBiasResults}
\title{plot Transcription and Replication strand bias results}
\usage{
plotBiasResults(biasResObj, filename = NULL, addToTitle = NULL)
}
\arguments{
\item{biasResObj}{R list object containing the results from the sampleStrandBias function call}

\item{filename}{pdf file name to save the plots to file}

\item{addToTitle}{text to be added in each plot title, useful for example to add the name of a sample}
}
\description{
This function plots the transcription/replication strand bias results obtained
using the sampleStrandBias function.
}
