% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationContext.R
\name{plotContextSNV}
\alias{plotContextSNV}
\title{plot SNV context}
\usage{
plotContextSNV(data_table, outfile = NULL, main = "", xlab = "")
}
\arguments{
\item{data_table}{table with counts of nucleotides at the 5' and 3' of a mutation, obtained from calling the findContextSNV function}

\item{outfile}{if specified the plot will be save to a pdf file, use extension .pdf}

\item{main}{optional plot title}

\item{xlab}{optional x label}
}
\description{
Plot the result table obtained from calling the findContextSNV function
}
