% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignaturesCompact.R
\name{plotDNVSignaturesCompact}
\alias{plotDNVSignaturesCompact}
\title{Plot Dinucleotide Variant Signatures or Catalogues, Compact plot}
\usage{
plotDNVSignaturesCompact(
  signature_data_matrix,
  output_file = NULL,
  plot_sum = FALSE,
  overall_title = "",
  add_to_titles = NULL,
  mar = NULL,
  cex.title = 1,
  cex.axistext = 1,
  howManyInOnePage = 100,
  ncolumns = 1
)
}
\arguments{
\item{signature_data_matrix}{matrix of signatures or catalogues, signatures as columns and channels as rows. You can use either Zou or Alexandrov Style and the function will recognise the rownames and plot accordingly}

\item{output_file}{set output file, should end with ".jpg" of ".pdf". If output_file==null, output will not be to a file, but will still run the plot functions. The option output_file==null can be used to add this plot to a larger output file.}

\item{plot_sum}{whether the sum of the channels should be plotted. If plotting signatures this should be FALSE, but if plotting sample catalogues, this can be set to TRUE to display the number of mutations in each sample.}

\item{overall_title}{set the overall title of the plot}

\item{add_to_titles}{text to be added to the titles of each catalogue plot}

\item{mar}{set the margin of the plot}

\item{howManyInOnePage}{how many signatures or catalogues should be plotted on one page. Multiple pages are plotted if more signatures/catalogues to plot have been requested}

\item{ncolumns}{how many columns should be used to arrange the signatures/catalogues to plot}
}
\description{
Function to plot one or more DNV signatures or catalogues.  Minimal compact plot without labels and axis.
}
