% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{plotFit}
\alias{plotFit}
\title{Plot the results from the Fit function}
\usage{
plotFit(fitObj, outdir = "", samplesInSubdir = TRUE)
}
\arguments{
\item{fitObj}{object obtained from the Fit function}

\item{outdir}{output directory where the results should be saved/plotted}

\item{samplesInSubdir}{if TRUE, move the sample files to a subdirectory named "samples"}
}
\description{
Plotting of the results obtained with the Fit function. Output adapts based on the options used in the Fit function
}
\examples{
res <- Fit(catalogues,getOrganSignatures("Breast"))
plotFit(res,"results/")
}
