% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{plotFitMS}
\alias{plotFitMS}
\title{Plot the results from the FitMS function}
\usage{
plotFitMS(fitMSobj, outdir = "")
}
\arguments{
\item{outdir}{output directory where the results should be saved/plotted}

\item{fitObj}{object obtained from the FitMS function}
}
\description{
Plotting of the results obtained with the FitMS function. Output adapts based on the options used in the FitMS function
}
\examples{
res <- FitMS(catalogues,"Breast")
plotFitMS(res,"results/")
}
