% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{plotFitResults}
\alias{plotFitResults}
\title{Plot the results from the Fit or FitMS function}
\usage{
plotFitResults(fitObj, outdir = "")
}
\arguments{
\item{fitObj}{object obtained from the Fit or FitMS function}

\item{outdir}{output directory where the results should be saved/plotted}
}
\description{
Plotting of the results obtained with the Fit or FitMS function. Output adapts based on the options used in during fitting.
You can use this plot function instead of using plotFit or plotFitMS. The function plotFitResults will infer whether Fit or FitMS
was used and use the appropriate plot function.
}
\examples{
res <- Fit(catalogues,getOrganSignatures("Breast"))
plotFitResults(res,"results/")
}
