% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{plotGenericSignatures}
\alias{plotGenericSignatures}
\title{Plot Generic Signatures or Catalogues}
\usage{
plotGenericSignatures(signature_data_matrix, output_file = NULL,
  plot_sum = TRUE, overall_title = "", mar = NULL)
}
\arguments{
\item{signature_data_matrix}{matrix of signatures, signatures as columns and channels as rows}

\item{output_file}{set output file, should end with ".jpg". If output_file==null, output will not be to a file, but will still run the plot functions. The option output_file==null can be used to add this plot to a larger output file.}

\item{plot_sum}{whether the sum of the channels should be plotted. If plotting signatures this should be FALSE, but if plotting sample catalogues, this can be set to TRUE to display the number of mutations in each sample.}

\item{overall_title}{set the overall title of the plot}

\item{mar}{set the option par(mar=mar)}
}
\description{
Function to plot one or more signatures or catalogues with an arbitrary number of channels. Channel names will not be plotted and all channels will be plotted as bars of the same colour.
}
