% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandBiasLib.R
\name{plotStrandBiasResults}
\alias{plotStrandBiasResults}
\title{plot Transcription and Replication strand bias results}
\usage{
plotStrandBiasResults(
  biasResObj,
  filename = NULL,
  addToTitle = NULL,
  pointsize = 12,
  textscaling = 1
)
}
\arguments{
\item{biasResObj}{R list object containing the results from the sampleStrandBias function call}

\item{filename}{pdf file name to save the plots to file}

\item{addToTitle}{text to be added in each plot title, useful for example to add the name of a sample}

\item{pointsize}{change the pointsize of the plot}

\item{textscaling}{change the scaling of some text}
}
\description{
This function plots the transcription/replication strand bias results obtained
using the sampleStrandBias function, or the combineStrandBiasResults function.
}
