% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{readTable}
\alias{readTable}
\title{readTable}
\usage{
readTable(file)
}
\arguments{
\item{file}{name of the plain text file to read}
}
\description{
Utility function for simple read table with the following parameters: (sep = "\\t",check.names = FALSE,header = TRUE,stringsAsFactors = FALSE).
}
