% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(m1, m2)
}
\arguments{
\item{m1}{first matrix to compare}

\item{m2}{second matrix to compare}
}
\value{
root mean squared error
}
\description{
Function to compute the root mean squared error between two matrices.
}
