% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandBiasLib.R
\name{sampleStrandBias}
\alias{sampleStrandBias}
\title{Transcription and Replication strand bias for a sample}
\usage{
sampleStrandBias(snv_table, genomev = "hg19")
}
\arguments{
\item{snv_table}{data frame with subs from a single sample and the following minimal columns: chr, position, REF, ALT.}

\item{genome.v}{either "hg38" (will load BSgenome.Hsapiens.NCBI.GRCh38), "hg19" (will load BSgenome.Hsapiens.UCSC.hg19), mm10 (will load BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10) or canFam3 (will load BSgenome.Cfamiliaris.UCSC.canFam3::BSgenome.Cfamiliaris.UCSC.canFam3)}
}
\value{
strand bias results
}
\description{
This function computes the transcription/replication strand bias of a given
set of single base substitutions, typically from a single sample. The function
returns the bias of the mutations with and without the trinucleotide context.
}
