% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCatalogueSimulationLib.R
\name{saveSimulatedData}
\alias{saveSimulatedData}
\title{Save simulated data}
\usage{
saveSimulatedData(simObj, simDir, saveAsPlainText = T)
}
\arguments{
\item{simObj}{object returned by the simpleCatalogueSimulation function}

\item{simDir}{directory where the simulated data should be stored}

\item{saveAsPlainText}{if TRUE data will be saved as plain text, in tab separated table format, while if FALSE data will be saved as R object.}
}
\description{
Function that can be used to save to file a dataset simulated with the
simpleCatalogueSimulation function.
}
