% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFitLib.R
\name{SignatureFit}
\alias{SignatureFit}
\title{Mutational Signatures Fit (deprecated)}
\usage{
SignatureFit(
  cat,
  signature_data_matrix,
  method = "KLD",
  bf_method = "CosSim",
  alpha = -1,
  doRound = TRUE,
  verbose = TRUE,
  n_sa_iter = 500,
  showDeprecated = TRUE
)
}
\arguments{
\item{cat}{catalogue matrix, patients as columns, channels as rows}

\item{signature_data_matrix}{signatures, signatures as columns, channels as rows}

\item{method}{KLD or NNLS or SA}

\item{bf_method}{bleeding filter method, one of KLD or CosSim, only if bleeding filter is used (alpha>-1)}

\item{alpha}{set alpha to -1 to avoid Bleeding Filter}

\item{doRound}{round the exposures to the closest integer}

\item{verbose}{use FALSE to suppress messages}

\item{n_sa_iter}{set max Simulated Annealing iterations if method==SA}

\item{showDeprecated}{set to FALSE to switch off the deprecated warning messsage}
}
\value{
returns the activities/exposures of the signatures in the given sample
}
\description{
This function is deprecated. You can still use it, but we advise to use the function Fit instead, 
which provides a unified interface for basic signature fit with/without bootstrap.
Fit a given set of mutational signatures into mutational catalogues to extimate the activty/exposure of each of the given signatures in the catalogues.
}
\examples{
res.exp <- SignatureFit(catalogues,signature_data_matrix)
}
\keyword{fit}
\keyword{mutational}
\keyword{signatures}
