% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCatalogueSimulationLib.R
\name{simpleCatalogueSimulation}
\alias{simpleCatalogueSimulation}
\title{simpleCatalogueSimulation}
\usage{
simpleCatalogueSimulation(
  commonSignatures,
  rareSignatures = NULL,
  ngenomes = 50,
  ncommon_signatures_per_sample = 5,
  nsamplesWrareSigs = 10,
  maxRareSigsPerSample = 1,
  min_mutations = 1000,
  max_mutations = 50000
)
}
\arguments{
\item{commonSignatures}{mutational signatures matrix with the signatures to be considered common}

\item{rareSignatures}{mutational signatures matrix with the signatures to be considered rare, can be omitted}

\item{ngenomes}{how many mutational catalogues should be simulated}

\item{ncommon_signatures_per_sample}{how many common signatures should be present in each sample}

\item{nsamplesWrareSigs}{how many samples should have rare signatures}

\item{maxRareSigsPerSample}{the maxiumum number of rare signature that can be present in a sample}

\item{min_mutations}{minimum number of mutations in a sample}

\item{max_mutations}{maximum number of mutations in a sample}
}
\value{
catalogues, signatures and exposures of simulated data
}
\description{
This is a simple function that produces simulated mutational catalgoues and exposures, given a set of mutational signatures.
Each catalogue is simulated as follows. The total number of mutations in a catalogue is sampled uniformely in the log space,
and the results exponetiated, so that most samples have total mutations closer to min_mutations than max_mutations.
exposures for common signatures and rare signatures (if any) are then sampled uniformely, and the result scaled so that the
sum of the mutations corresponds to the number sampled initially.
}
