% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DinucleotideVariantsLib.R
\name{snvTabToDNVcatalogue}
\alias{snvTabToDNVcatalogue}
\title{Build a Dinucleotide Variants Catalogue from SNVs}
\usage{
snvTabToDNVcatalogue(snvtab)
}
\arguments{
\item{snvtab}{requires columns Sample, Chrom, Pos, Ref, Alt, with Ref and Alt of length 1}
}
\value{
list of DNVs and DNV catalogue
}
\description{
This function takes as input a list of single nucleotide variants (SNVs),
and computes a list of dinucleotide variants (DNVs) finding which SNVs
are next to each other. It then returns the annotated DNVs and the DNV catalogues.
The catalogues are in Zou's style.
}
\references{
J. E. Kucab, X. Zou, S. Morganella, M. Joel, A. S. Nanda, E. Nagy, C. Gomez, A. Degasperi, R. Harris, S. P. Jackson, V. M. Arlt, D. H. Phillips, S. Nik-Zainal. A Compendium of Mutational Signatures of Environmental Agents. Cell, https://doi.org/10.1016/j.cell.2019.03.001, 2019.
}
