% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{sortCatalogue}
\alias{sortCatalogue}
\title{Sort 96-channel Substitution Catalogues}
\usage{
sortCatalogue(cat)
}
\arguments{
\item{cat}{catalogues matrix where each column is a catalogue and each row is a channel. Rownames should be set as the name of the channel in the format 5' base[Normal base>Tumour base]3' base, for example A[C>A]A.}
}
\value{
ordered catalogue
}
\description{
This function sorts a matrix of 96-channel Substitution Catalogues,
so that the channels (rows) are in the correct order.
where each column is a catalogue and each row is a channel.
rownames should be set as the name of the channel in the format
5' base[Normal base>Tumour base]3' base, for example A[C>A]A.
}
