% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DinucleotideVariantsLib.R
\name{tabToDNVcatalogue}
\alias{tabToDNVcatalogue}
\title{TAB to DNV catalogue}
\usage{
tabToDNVcatalogue(muttable)
}
\arguments{
\item{muttable}{data frame with the mutations, formatted with the following column names: chr, position, REF, ALT.}
}
\value{
returns the DNV catalogue for the given sample and mutation list
}
\description{
Convert a data frame containing SNVs and DNVs to a DNV catalogue. The data frame should containt the SNVs and/or DNVs of a single sample. Two SNVs that are next to each other will be combined into a DNV.
The data frame should have the following columns: chr, position, REF, ALT.
}
\examples{
muttable <- readTable("mutations.tsv")
res <- tabToSNVcatalogue(muttable)
}
\keyword{DNV}
\keyword{catalogue}
