% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabToIndelsClassification.R
\name{tabToIndelsClassification}
\alias{tabToIndelsClassification}
\title{tab to Indels Classification}
\usage{
tabToIndelsClassification(indel.data, sampleID, genome.v = "hg19")
}
\arguments{
\item{indel.data}{dataframe with indels from a single sample and the following minimal columns: chr, position, REF, ALT.}

\item{sampleID}{name of the sample}

\item{genome.v}{version of the genome to be used to look up the context of the indel, either "hg19", "hg38" or "mm10"}
}
\value{
the function returns a list with elements "indels_classified", which is a table with the indels and their classification, and "count_proportion", which is a summary of the count of indels and their proportion
}
\description{
Convert a dataframe containing Indels into a data frame where each indel is classified as repet-mediated, Microhomology-mediated or other. A summary of the count of indels (deletions and insertions) and their proportion with respect to the total is also provided.
}
\examples{
res <- tabToIndelsClassification(indel.data,"testSample","hg19")
}
