% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinucleotideVariantsLib.R
\name{tnvTabToTNVcatalogue}
\alias{tnvTabToTNVcatalogue}
\title{Build a Trinucleotide Variants Catalogue from TNVs list}
\usage{
tnvTabToTNVcatalogue(tnvtab)
}
\arguments{
\item{tnvtab}{requires columns Sample, Chrom, Pos, Ref, Alt, with Ref and Alt of length 3}
}
\value{
list of annotated TNVs and TNV catalogue
}
\description{
This function takes as input a list of trinucleotide variants (TNVs). It then
annotates the TNVs and computes the TNV catalogues.
The trinucleotide mutations considered are only those where all three bases change,
i.e. TTT>AAA but not TTT>ATA.
}
