% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceEvaluationLib.R
\name{updateSigNamesWithMatchTable}
\alias{updateSigNamesWithMatchTable}
\title{update signature names using a match table}
\usage{
updateSigNamesWithMatchTable(signames, matchTable)
}
\arguments{
\item{signames}{vector of signature names, typically estimated signatures, that need to be replaced}

\item{matchTable}{table with matched true and estimated signature names}
}
\value{
updated signames vector
}
\description{
This function is part of the workflow of simulating a dataset, testing a
signature extraction and evaluating the signatures/exposures accuracy.
The function evaluatePerformanceSignatureSimilarity returns a table
matching true signatures with estimated signatures. The updateSigNamesWithMatchTable
function can be used with the match table to replace the names of the
estimated signatures with those of the true signatures, for example in the
matrix of estimated exposures. See example below.
}
\examples{
sigsPerf <- evaluatePerformanceSignatureSimilarity(true_signatures,
                                                   estimated_signatures)
updatedNames <- updateSigNamesWithMatchTable(signames,
                                             sigsPerf$matchTable)

}
