% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DinucleotideVariantsLib.R
\name{vcfToDNVcatalogue}
\alias{vcfToDNVcatalogue}
\title{VCF to DNV catalogue}
\usage{
vcfToDNVcatalogue(vcfFilename, genome.v = "hg19")
}
\arguments{
\item{vcfFilename}{path to input VCF (file must be tabix indexed)}

\item{genome.v}{either "hg38" (will load BSgenome.Hsapiens.UCSC.hg38), "hg19" (will load BSgenome.Hsapiens.1000genomes.hs37d5), mm10 (will load BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10) or canFam3 (will load BSgenome.Cfamiliaris.UCSC.canFam3::BSgenome.Cfamiliaris.UCSC.canFam3)}
}
\value{
returns the DNV catalogue for the given sample and mutation list
}
\description{
Convert a vcf file containing SNVs and DNVs to a DNV catalogue. The VCF file should containt the SNVs and/or DNVs of a single sample. Two SNVs that are next to each other will be combined into a DNV.
}
\examples{
file_muts <- "mutations.vcf"
res <- vcfToSNVcatalogue(file_muts,genome.v = "hg38")
}
\keyword{DNV}
\keyword{vcf}
