% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndelSignaturesLib.R
\name{vcfToIndelsCatalogue89}
\alias{vcfToIndelsCatalogue89}
\title{vcf to Indels Catalogue (89 channels)}
\usage{
vcfToIndelsCatalogue89(
  indelsVcfFile,
  sampleName,
  useHighSpecificityFilter = TRUE,
  genome.v
)
}
\arguments{
\item{indelsVcfFile}{filename of the vcf file containing only indels}

\item{sampleName}{name of the sample to process}

\item{useHighSpecificityFilter}{recommended TRUE. Filters out indels in highly repeated regions (>= 10 repeats) and indels longer than 100 bps}

\item{genome.v}{version of the genome to be used to look up the context of the indel, either "hg19", "hg38", "mm10", "canFam3"}
}
\value{
the function returns classified indels, as well as the indel catalogue
}
\description{
Convert a dataframe containing Indels into a data frame where each indel is classified as repet-mediated, Microhomology-mediated or other. A summary of the count of indels (deletions and insertions) and their proportion with respect to the total is also provided.
}
\examples{
res <- vcfToIndelsCatalogue89(indel.data,"sample","hg19")
}
