% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfToIndelsClassification.R
\name{vcfToIndelsClassification}
\alias{vcfToIndelsClassification}
\title{VCF to Indels Classification}
\usage{
vcfToIndelsClassification(indelsVCF.file, sampleID, genome.v = "hg19")
}
\arguments{
\item{indelsVCF.file}{path to input VCF (file must be tabix indexed). This file should have been already filtered for the final indels sets to be used in the analysis.}

\item{sampleID}{name of the sample}

\item{genome.v}{version of the genome to be used to look up the context of the indel, either "hg19", "hg38", "mm10" or "canFam3"}
}
\value{
the function returns a list with elements "indels_classified", which is a table with the indels and their classification, and "count_proportion", which is a summary of the count of indels and their proportion
}
\description{
Convert a VCF file containing Indels into a data frame where each indel is classified as repet-mediated, Microhomology-mediated or other. A summary of the count of indels (deletions and insertions) and their proportion with respect to the total is also provided.
}
\examples{
res <- vcfToIndelsClassification("test.indel.vcf.gz","testSample","hg19")
}
