% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfToSNVcatalogue.R
\name{vcfToSNVcatalogue}
\alias{vcfToSNVcatalogue}
\title{VCF to SNV catalogue}
\usage{
vcfToSNVcatalogue(vcfFilename, genome.v = "hg19")
}
\arguments{
\item{vcfFilename}{name of the VCF file to read from}

\item{genome.v}{either "hg38" (will load BSgenome.Hsapiens.NCBI.GRCh38) or "hg19" (will load BSgenome.Hsapiens.UCSC.hg19)}
}
\value{
returns the SNV catalogue for the given sample
}
\description{
Convert a vcf file containing SNV to SNV 96 channel trinuclotide context catalogue. The VCF file should containt the SNV of a single sample.
}
\examples{
file_subs <- "subs.vcf"
res <- vcfToSNVcatalogue(file_subs,genome.v = "hg38")
}
\keyword{SNV}
\keyword{vcf}
