% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{writeTable}
\alias{writeTable}
\title{writeTable}
\usage{
writeTable(t, file)
}
\arguments{
\item{t}{R table or matrix}

\item{file}{name of the output plain text file}
}
\description{
Utility function for simple write table with the following parameters: (sep = "\\t",quote = FALSE,row.names = TRUE,col.names = TRUE).
}
