% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascatToHRDLOH.R
\name{ascatToHRDLOH}
\alias{ascatToHRDLOH}
\title{HRD-LOH index from ASCAT}
\usage{
ascatToHRDLOH(ascat.data, SAMPLE.ID)
}
\arguments{
\item{ascat.data}{data frame with the following columns: 'seg_no', 'Chromosome', 'chromStart', 'chromEnd', 'total.copy.number.inNormal', 'minor.copy.number.inNormal', 'total.copy.number.inTumour', 'minor.copy.number.inTumour'}

\item{SAMPLE.ID}{sample name}
}
\description{
Compute the HRD-LOH index starting from a data fram (possibly loaded from an ASCAT output file)
with the following columns: 'seg_no', 'Chromosome', 'chromStart', 'chromEnd', 'total.copy.number.inNormal', 
'minor.copy.number.inNormal', 'total.copy.number.inTumour', 'minor.copy.number.inTumour'
}
