% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascatToHRDLOH.R
\name{ascatToHRDLOH}
\alias{ascatToHRDLOH}
\title{HRD-LOH index from ASCAT}
\usage{
ascatToHRDLOH(ascat.data, SAMPLE.ID)
}
\arguments{
\item{ascat.data}{data frame with the following columns: 'seg_no', 'Chromosome', 'chromStart', 'chromEnd', 'total.copy.number.inNormal', 'minor.copy.number.inNormal', 'total.copy.number.inTumour', 'minor.copy.number.inTumour'}

\item{SAMPLE.ID}{sample name}
}
\value{
returns the HRD-LOH index
}
\description{
Compute the HRD-LOH index starting from a data fram (possibly loaded from an ASCAT output file)
with the following columns: 'seg_no', 'Chromosome', 'chromStart', 'chromEnd', 'total.copy.number.inNormal', 
'minor.copy.number.inNormal', 'total.copy.number.inTumour', 'minor.copy.number.inTumour'
}
\examples{
ascat.data <- read.table("ascat.scv",sep=",",header=TRUE)
hrd_index <- ascatToHRDLOH(ascat.df,"test_sample")
}
\references{
Abkevich, V., Timms, K. M., Hennessy, B. T., Potter, J., Carey, M. S., Meyer, L. a., ... Lanchbury, J. S. (2012). Patterns of genomic loss of heterozygosity predict homologous recombination repair defects in epithelial ovarian cancer. British Journal of Cancer, 107(10), 1776–82. https://doi.org/10.1038/bjc.2012.451
}
