% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedpeToRearrCatalogue.R
\name{bedpeToRearrCatalogue}
\alias{bedpeToRearrCatalogue}
\title{bedpeToRearrCatalogue}
\usage{
bedpeToRearrCatalogue(sv_bedpe)
}
\arguments{
\item{sv_bedpe}{data frame BEDPE as described above}
}
\value{
returns a list with 1) the rearrangement catalogue for the given sample and 2) the annotated bedpe for the given sample
}
\description{
This function converts a data frame BEDPE into a rearrangement catalogue, 
you should pass rearrangements of only one sample, and one rearrangement for each paired-end mates.
The BEDPE data fram should contain the following columns: "chrom1", "start1", "end1", "chrom2", "start2", "end2" and "sample" (sample name). 
In addition, either two columns indicating the strands of the mates, "strand1" (+ or -) and "strand2" (+ or -), or one column indicating the structural variant class, "svclass": translocation, inversion, deletion, tandem-duplication.
}
\details{
The column "svclass" should correspon to (Sanger BRASS convention): (strand1/strand2)

inversion (+/-), if mates on the same chromosome

inversion (-/+), if mates on the same chromosome

deletion (+/+), if mates on the same chromosome

tandem-duplication (-/-), if mates on the same chromosome

translocation, if mates are on different chromosomes
}
\examples{
vcf_sv_file.bedpe <- "sample.bedpe"
sv_bedpe <- read.table(vcf_sv_file.bedpe,sep = "\\t",header = TRUE,
                     stringsAsFactors = FALSE,check.names = FALSE)
#build a catalogue from the bedpe file
res.cat <- bedpeToRearrCatalogue(sv_bedpe)
}
\keyword{bedpe}
\keyword{rearrangement}
