% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{fitMerge}
\alias{fitMerge}
\title{Selecting and merging signature fit results from Multi-Step signature fit}
\usage{
fitMerge(resObj, forceRareSigChoice = NULL)
}
\arguments{
\item{resObj}{result object obtained from the FitMS function}

\item{forceRareSigChoice}{if NULL this function will select the rare signature candidate with the highest associated cosine similarity.
If no rare signature is found, then the solution with only the common signatures is selected.
To select specific candidates, specify them in the forceRareSigChoice list object, in the form forceRareSigChoice[["sample_name"]] <- "rareSigName".
To select the solution with only the common signatures for a sample use forceRareSigChoice[["sample_name"]] <- "common"}
}
\value{
returns the updated resObj object with updated exposures and rareSigChoice objects.
If bootstrap was used, bootstraps of selected solutions can be found in the variable bootstrap_exposures_samples
}
\description{
This function is used to select candidate rare signature solutions from a result object obtained
using the multi-step signature fit FitMS function.
}
\details{
When running FitMS, some samples may have multiple candidate rare signatures or rare signature combinations that
fit the sample reasonably well. This function selects the best rare signature for each sample based on cosine similarity and returns
an updated object with a summary exposures composed by each selected fit solution for each sample.

The choise of rare signature can be changed using the parameter forceRareSigChoice.
}
