% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeChart.R
\name{genomeChart}
\alias{genomeChart}
\title{genomeChart}
\usage{
genomeChart(
  outfilename,
  sample_name,
  SNV_vcf_file = NULL,
  SNV_tab_file = NULL,
  Indels_vcf_file = NULL,
  Indels_tab_file = NULL,
  CNV_tab_file = NULL,
  SV_bedpe_file = NULL,
  plot_title = NULL,
  runKataegis = TRUE,
  genome.v = "hg19",
  debug = FALSE
)
}
\arguments{
\item{outfilename}{file where the figure should be plotted. This also determines the file type of the output, use either pdf (recommended) or png}

\item{sample_name}{name of sample}

\item{SNV_vcf_file}{name of the vcf file containing the SNVs}

\item{SNV_tab_file}{name of the tab separated file containing the SNVs. Column names should be: chr, position, REF and ALT. If SNV_vcf_file is also specified, these variants will be ignored and the variants in the vcf file will be used instead.}

\item{Indels_vcf_file}{name of the vcf file containing the Indels}

\item{Indels_tab_file}{name of the tab separated file containing the Indels. Column names should be: chr, position, REF and ALT. If Indels_vcf_file is also specified, these variants will be ignored and the variants in the vcf file will be used instead.}

\item{CNV_tab_file}{name of the tab separated file containing the CNVs. Column names should be: 'seg_no', 'Chromosome', 'chromStart', 'chromEnd', 'total.copy.number.inNormal', 'minor.copy.number.inNormal', 'total.copy.number.inTumour', 'minor.copy.number.inTumour'}

\item{SV_bedpe_file}{name of the tab separated bedpe file containing the SVs. The file should contain a rearrangement for each row (two breakpoint positions should be on one row as determined by a pair of mates of paired-end sequencing) and should already be filtered according to the user preference, as all rearrangements in the file will be used and no filter will be applied. The files should contain a header in the first line with the following columns: "chrom1", "start1", "end1", "chrom2", "start2", "end2" and "sample" (sample name). In addition, either two columns indicating the strands of the mates, "strand1" (+ or -) and "strand2" (+ or -), or one column indicating the structural variant class, "svclass": translocation, inversion, deletion, tandem-duplication. The column "svclass" should correspond to (Sanger BRASS convention): inversion (strands +/- or -/+ and mates on the same chromosome), deletion (strands +/+ and mates on the same chromosome), tandem-duplication (strands -/- and mates on the same chromosome), translocation (mates are on different chromosomes). In addition, columns 'non-template'    and 'micro-homology' can be specified, including the non-templated insertion or micro-homology deletion sequence, which will be used to build an SV junctions catalogue.}

\item{plot_title}{title of the plot. If NULL, then the sample_name will be used as title. Use "", the empty string, to have no title.}

\item{runKataegis}{whether or not to run the kataegis algorithm (default is TRUE)}

\item{genome.v}{genome version to use, either hg19 or hg38}

\item{debug}{if TRUE, show debug guidelines and grid when plotting (default is FALSE)}
}
\value{
all computed results, like catalogues and clustering regions, will be returned
}
\description{
Plotting of somatic variants using the circlize R package. Somatic variants
plotted are single nucleotide variants (SNVs), Indels, copy number variants
(CNVs) and structural variants (SVs or rearrangements). In addition to the
circle visualisation of the variants, other data are plotted, such as SNV and
SV mutational catalogues, indels classification summary counts and a copy number
plot indicating estimates of total and minor allele copy numbers.
}
