% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{getReferenceSignatures}
\alias{getReferenceSignatures}
\title{getReferenceSignatures}
\usage{
getReferenceSignatures(version = "latest", typemut = "subs", verbose = TRUE)
}
\arguments{
\item{version}{version "1" includes subs or rearr (ICGC cohort) reference signatures from Degasperi et al. 2020, while version "2" includes the improved subs reference signatures, and the new DNV signatures. Set to "latest" to get the latest signature available for a given mutation type.}

\item{typemut}{either subs, DNV or rearr}
}
\value{
reference signatures matrix
}
\description{
This function returns the reference signatures for a given mutation type as defined in Degasperi et al. 2020 Nat Cancer paper,
and Degasperi et al. 2022 Science paper.
}
\references{
A. Degasperi, T. D. Amarante, J. Czarnecki, S. Shooter, X. Zou, D. Glodzik, S. Morganella, A. S. Nanda, C. Badja, G. Koh, S. E. Momen, I. Georgakopoulos-Soares, J. M. L. Dias, J. Young, Y. Memari, H. Davies, S. Nik-Zainal. A practical framework and online tool for mutational signature analyses show intertissue variation and driver dependencies, Nature Cancer, https://doi.org/10.1038/s43018-020-0027-5, 2020.
}
