% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures_utils.R
\name{getSignaturesForFitting}
\alias{getSignaturesForFitting}
\title{getSignaturesForFitting}
\usage{
getSignaturesForFitting(organ, typemut = "subs", tier = "T2", verbose = TRUE)
}
\arguments{
\item{organ}{organs available are: "Biliary", "Bladder", "Bone_SoftTissue", "Breast", "CNS", "Colorectal", "Esophagus", "Head_neck", "Kidney", "Liver", "Lung", "Lymphoid", "Myeloid", "NET", "Oral_Oropharyngeal", "Ovary", "Pancreas", "Prostate", "Skin", "Stomach", "Uterus"}

\item{typemut}{only subs supported at the moment}

\item{tier}{either "T1" or "T2". "T2" is default. For each organ we provide two lists of rare signatures that can be used. Tier 1 (T1) are rare signatures
that were observed in the requested organ. The problem with T1 is that it may be that a signature is not observed simply because there were not enough samples for a certain organ in the particular
dataset that was used to extract the signatures. So in general we advise to use Tier 2 (T2) signatures, which extend the rare signature to a wider number of rare signatures.
More specifically, T2 includes all the reference signatures that were observed as rare in the specified organ and also reference signatures that were observed as rare in other at least two organs.}
}
\value{
list of signatures matrix
}
\description{
Given a tissue type/organ, his function returns the common and rare mutational signatures
to be used with FitMS, as suggested in Degasperi et al. 2022, Science paper.
}
