% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDetect.R
\name{plot_HRDetect_BootstrapScores}
\alias{plot_HRDetect_BootstrapScores}
\title{plot_HRDetect_BootstrapScores}
\usage{
plot_HRDetect_BootstrapScores(
  outdir,
  hrdetect_res,
  main = "Distribution of HRDetect scores",
  mar = NULL,
  pwidth = NULL,
  pheight = NULL,
  pointsize = 12
)
}
\arguments{
\item{outdir}{output directory for the plot}

\item{hrdetect_res}{output of HRDetect_pipeline, ran with bootstrapHRDetectScores=TRUE}

\item{main}{plot title. If not specified: "Distribution of HRDetect scores"}

\item{mar}{plot margin. If not specified, this will be dynamically estimated based on sample names length}

\item{pwidth}{plot width in pixels. If not specified, this will be dynamically estimated based on the number of samples}

\item{pheight}{plot height in pixels. If not specified, this will be dynamically estimated based on sample names length}

\item{pointsize}{plot resolution. If not specified: 12}
}
\description{
Overall plot of scores obtained from HRDetect with bootstrap. Plot will be in pdf format.
}
