% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Plot Matrix}
\usage{
plotMatrix(
  dataMatrix,
  output_file = NULL,
  thresholdMark = NULL,
  ndigitsafterzero = 2,
  cex.numbers = 0.7,
  circlesColBasic = "#A1CAF1",
  circlesColHighlight = "#F6A600"
)
}
\arguments{
\item{dataMatrix}{a data matrix or data frame}

\item{output_file}{if an output file name is given (must be pdf), the matrix will be plotted to file}

\item{thresholdMark}{threshold for using a different colour to highlight entries above a threshold}

\item{ndigitsafterzero}{specify how many digits after the zero should be used to show the actual numbers}

\item{cex.numbers}{scale the text used for the numbers in the matrix}

\item{circlesColBasic}{colour used for the circles}

\item{circlesColHighlight}{colour used for the circles that pass the thresholdMark}
}
\description{
This function plots a matrix of values. Data is visualised as circles scaled with respect to the
largest value in the matrix, with the actual number shown. If thresholdMark is specified, then a
different colour is used for entries that are equal or above the threshold.
}
