% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFitLib.R
\name{SignatureFit_withBootstrap}
\alias{SignatureFit_withBootstrap}
\title{Mutational Signatures Fit with Bootstrap (deprecated)}
\usage{
SignatureFit_withBootstrap(
  cat,
  signature_data_matrix,
  nboot = 100,
  exposureFilterType = "fixedThreshold",
  giniThresholdScaling = 10,
  threshold_percent = 5,
  threshold_p.value = 0.05,
  method = "KLD",
  bf_method = "CosSim",
  alpha = -1,
  verbose = TRUE,
  doRound = FALSE,
  nparallel = 1,
  n_sa_iter = 500,
  randomSeed = NULL,
  showDeprecated = TRUE
)
}
\arguments{
\item{cat}{catalogue matrix, patients as columns, channels as rows}

\item{signature_data_matrix}{signatures, signatures as columns, channels as rows}

\item{nboot}{number of bootstraps to use, more bootstraps more accurate results}

\item{exposureFilterType}{use either fixedThreshold or giniScaledThreshold. When using fixedThreshold, exposures will be removed based on a fixed percentage with respect to the total number of mutations (threshold_percent will be used). When using giniScaledThreshold each signature will used a different threshold calculated as (1-Gini(signature))*giniThresholdScaling}

\item{giniThresholdScaling}{scaling factor for the threshold type giniScaledThreshold, which is based on the Gini score of a signature}

\item{threshold_percent}{threshold in percentage of total mutations in a sample, only exposures larger than threshold are considered}

\item{threshold_p.value}{p-value to determine whether an exposure is above the threshold_percent. In other words, this is the empirical probability that the exposure is lower than the threshold}

\item{method}{KLD or NNLS or SA}

\item{bf_method}{bleeding filter method, one of KLD or CosSim, only if bleeding filter is used (alpha>-1)}

\item{alpha}{set alpha to -1 to avoid Bleeding Filter}

\item{verbose}{use FALSE to suppress messages}

\item{doRound}{round the exposures to the closest integer}

\item{nparallel}{to use parallel specify >1}

\item{n_sa_iter}{set max Simulated Annealing iterations if method==SA}

\item{randomSeed}{set an integer random seed}

\item{showDeprecated}{set to FALSE to switch off the deprecated warning messsage}
}
\value{
returns the activities/exposures of the signatures in the given sample and other information, such as p-values and exposures of individual bootstrap runs.
}
\description{
This function is deprecated. You can still use it, but we advise to use the function Fit instead, 
which provides a unified interface for basic signature fit with/without bootstrap.
Fit a given set of mutational signatures into mutational catalogues to extimate the 
activty/exposure of each of the given signatures in the catalogues. Implementation 
of method similar to Huang et al. 2017, Detecting presence of mutational signatures with 
confidence, which uses a bootstrap apporach to calculate the empirical probability 
of an exposure to be larger or equal to a given threshold (i.e. 5% of mutations of a sample). 
This probability can be used to decide which exposures to remove from the initial fit, 
thus increasing the sparsity of the exposures.
}
\examples{
res <- SignatureFit_withBootstrap(catalogues,signature_data_matrix)
}
\references{
Huang, X., Wojtowicz, D., & Przytycka, T. M. (2017). Detecting Presence Of Mutational Signatures In Cancer With Confidence. bioRxiv, (October). https://doi.org/10.1101/132597
}
\keyword{fit}
\keyword{mutational}
\keyword{signatures}
