% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfToSNVcatalogue.R
\name{vcfToSNVcatalogue}
\alias{vcfToSNVcatalogue}
\title{VCF to SNV catalogue}
\usage{
vcfToSNVcatalogue(vcfFilename, genome.v = "hg19")
}
\arguments{
\item{vcfFilename}{path to input VCF (file must be tabix indexed)}

\item{genome.v}{either "hg38" (will load BSgenome.Hsapiens.UCSC.hg38), "hg19" (will load BSgenome.Hsapiens.1000genomes.hs37d5), mm10 (will load BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10) or canFam3 (will load BSgenome.Cfamiliaris.UCSC.canFam3::BSgenome.Cfamiliaris.UCSC.canFam3)}
}
\value{
returns the SNV catalogue for the given sample
}
\description{
Convert a vcf file containing SNV to SNV 96 channel trinuclotide context catalogue. The VCF file should containt the SNV of a single sample.
}
\examples{
file_subs <- "subs.vcf"
res <- vcfToSNVcatalogue(file_subs,genome.v = "hg38")
}
\keyword{SNV}
\keyword{vcf}
