% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFitMultiStepLib.R
\name{writeFitResultsToJSON}
\alias{writeFitResultsToJSON}
\title{Write the results from the Fit or FitMS function to JSON}
\usage{
writeFitResultsToJSON(fitObj, compress = TRUE, filename = "fitData.json")
}
\arguments{
\item{fitObj}{object obtained from the Fit or FitMS function}

\item{compress}{if TRUE then compress the JSON file. This is recommended as a plain JSON text file can be quite large. Set to TRUE as default}

\item{filename}{output file name of the JSON file. If the parameter compress is TRUE then the ".zip" postfix will be added}
}
\description{
Writing of the results obtained with the Fit or FitMS function to JSON.
}
\examples{
res <- Fit(catalogues,getOrganSignatures("Breast"))
writeFitResultsToJSON(res,"results/")
}
